portage: Drop CONTENTS from sparse vdb layers
The `CONTENTS` file causes problems with interface libraries because it
contains the hashes of all the files we strip out of the interface
layer. This means that we still cache bust all reverse dependencies
when one of these files changes.

This will cause problems when building with 9999 ebuilds because the
find-missing-deps hook relies on the CONTENTS file. Since the
find-missing-deps script provides minimal value in the bazel context,
I plan on removing it from the hooks.

BUG=b:261863720
TEST=bazel test //bazel/portage/...

Change-Id: I669bcc0f0f13918616d4f8b699a1b34cca05c833
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5539333
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>