portage: Reusable deps: Add test packages (Alchemist integration tests)
The purpose of this CL is to add some packages that will cause diffs in
follow-up CLs that illustrate the effect of reusing package
installations from dependencies.

By itself, this CL contributes nothing new. But by defining these
packages in a separate CL, I hope to make the follow-up CLs easier to
review.

The important packages introduced here are reusabledeps-a and
reusabledeps-b. The former produces a reusable SDK that can be leveraged
by the latter. All other packages defined in this CL (pkg-00, ...,
pkg-15) are part of the dependency graph of reusabledeps-a and
reusabledeps-b, and do not attempt to illustrate any of the followup
changes by themselves.

BUG=b:342012804
TEST=Run "cargo test" inside alchemist dir

Change-Id: I36b60d28c9dfdb673f281af71d81df8defdda3d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5617781
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Leandro Lovisolo <lovisolo@google.com>