action_wrapper: Set up $TMPDIR on running actions
In an upcoming patch to support metacopy layers in durable trees, we
have to create metacopy directories in non-overlayfs mounts so as not
to hit the overlayfs lowerdir limit. Historically we avoided the similar
issue in the container crate by providing a "mutable base directory"
explicitly. However it's tedious to do the same with durable trees
because durable trees are used everywhere in Alchemy.

So it's time to reconsider how to provide temporary directories to our
actions.

This patch adds a flag "--temp-dir" to action_wrapper that instructs it
to set $TMPDIR before starting a program. It is a required option so it
is an error to forget to set the temporary directory. Starlark for every
action are updated to compute an appropriate temporary directory
(typically "tmp" directory in the package output directory). Also, most
users of the container crate are updated to not set the mutable base
directory explicitly and just rely on the default value.

BUG=b:340382187
TEST=BOARD=brya bazel build @portage//internal/packages/stage1/target/\
     host/portage-stable/virtual/os-headers:0-r2
TEST=BOARD=brya bazel run @portage//internal/packages/stage1/target/\
     host/portage-stable/virtual/os-headers:0-r2_debug

Change-Id: I316e2d7c5ca0d1da132137f51cf34bd6ede748b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5546714
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>