alchemist: Add supports_interface_libraries TOML support
This new TOML attribute can be used to declare if a package supports
building with interface libraries. I wanted to support both boolean
and string values on the attribute, so I had to open code the parsing
of the toml. This change only adds the new attribute, follow up CLs
will consume it.

BUG=b:261863720
TEST=cargo test

Change-Id: Ic227359abd3426db7977c969a80b1d46495844a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5555991
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>