alchemist: Clean up supports_interface_libraries API
When `supports_interface_libraries` was implemented, it only took the
last value that was found. This makes it a bit hard to reason about.
Instead AND the result of all the expressions. This way if one eclass
says it supports interface libraries and another says it doesn't,
we default to not supporting them.

I also refactored the code a bit so that we now parse the bash expr
when loading the TOML. I was able to remove the open-coded parser by
implementing Deserialize for BashExpr. This did require that I make the
TomlMetadata private since the TOML format differs from the public API.

This change also allows us to unit test the expression analysis logic.

BUG=b:261863720
TEST=cargo test

Change-Id: Ie4968c64601c730adc47b7577a0b5ab42270a5ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5567093
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>