create_interface_layer: Add --include flag
The --include flag can be used to specify files that we always want to
include into the interface layer. By default we strip all `.a` files,
but some packages use a hybrid approach where the bulk of code is
supplied by a `.so`, but some parts are supplied in a `.a.

There are also some packages that produce ASCII `.so` files, instead of
`.elf` files (i.e., /usr/lib64/libbase.so). We can't convert these using
llvm-ifs, so we can use the `--include` flag to just copy the file
instead of trying ti convert it into an interface library.

BUG=b:261863720
TEST=cargo test

Change-Id: I9ac702d2abf4cf2e6e6a779277bf446369dc1cae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5571253
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>