create_interface_layer: Handle more restrictive umask
Cloudtops run with umask 0027 while my desktop runs with 0022. This
means that when git checks out the files, they are missing the expected
`other` permissions bits. This change copies the group permission bits
to the other permissions so we land in an expected state.

BUG=b:261863720
TEST=cargo test

Change-Id: Ibfb2ca48544ff1affd674282cc8520b0b5f561eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5567089
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>