portage: Consolidate interface sdk_install_deps layer
This change makes it so that a single `sdk_install_deps` target can
handle building packages with and without interface libraries. We can do
this because the `sparse` layer is always used when "installing"
packages. When interface libraries are enabled, we just swap out the
"installed" content layer. This makes it so the ebuild is in control
of which layer types they want.

This allows us to drop the special `_interface_deps` target and keep the
original `_deps` name.

BUG=b:261863720
TEST=BOARD=amd64-generic bazel aquery --//bazel/portage:enable_interface_libraries @portage//internal/packages/stage2/target/board/portage-stable/app-emulation/qemu:9.0.0-r1
TEST=BOARD=amd64-generic bazel aquery --//bazel/portage:enable_interface_libraries @portage//internal/packages/stage2/target/board/portage-stable/app-emulation/qemu:9.0.0-r1_non_interface_test_package

Disallow-Looks-For-Green: True
Change-Id: I381279d8524120ae98512408f18d119619f5f21f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5598463
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>