workspace_root: fwsdk: Drop "nonhermetic" package
This package comes from opentitan, and currently isn't being used.
It's problematic as opentitan is a private repository, but we expect
fwsdk to work in a public checkout.  One way to support this is to
move it to it's own module extension, but I think we'd rather pull in
opentitan from subtool instead.

BUG=b:338458418
TEST=CQ

Change-Id: I370942bae262e33598f6de35da9a5ac9017536a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5753762
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jett Rink <jettrink@google.com>
Reviewed-by: Jett Rink <jettrink@google.com>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>